requirements.txt
grpcio==1.31.0
grpcio-tools==1.31.0
bigartm==0.9.2
click==7.0
pandas==1.0.5
zhon==1.1.5
pyyaml==5.3.1